requirements.txt
# Bibliotecas principais
pandas==2.0.3
numpy==1.24.3
scipy==1.10.1

# Machine Learning
scikit-learn==1.2.2
joblib==1.3.2

# Interface Web
streamlit==1.28.2
plotly==5.18.0

# Visualização
matplotlib==3.7.1
seaborn==0.12.2

# Utilitários
python-dotenv==1.0.0
requests==2.31.0
PyYAML==6.0.1

# Testes
pytest==7.4.3